Add bitmap dumping support

+++ include/pe_resource_loader.h
@@ -19,7 +19,7 @@ uint8_t * PeResourceLoader_GetString(PeResourceLoader * loader, uint32_t languag
 
 uint32_t * PeResourceLoader_GetBitmapLanguageIds(PeResourceLoader * loader, uint16_t * language_count);
 uint32_t * PeResourceLoader_GetBitmapIds(PeResourceLoader * loader, uint16_t * bitmap_count);
-uint8_t * PeResourceLoader_GetBitmap(PeResourceLoader * loader, uint32_t language_id, uint32_t bitmap_id, uint16_t * length);
+uint8_t * PeResourceLoader_GetBitmap(PeResourceLoader * loader, uint32_t language_id, uint32_t bitmap_id, uint32_t * length);
 
 
 // Possible languages
